configure visual e2e testing
e.g. if pollen calendar looks drastically different, then test will fail.

for this example, the way to solve it would fix bug that was ruining calendar OR if it is an improvement to calendar, update the snapshots in `cypress/snapshots/base/<testfilename>/<snapshotname>`

followed the `readme` of the package to setup ☝️
https://github.com/cypress-visual-regression/cypress-visual-regression