bug 🐛🐛: using date filter makes charts have weird offset 😬🤢🤢
when changing the date period using the date filter, the last bar or line points on the chart is offsetted by 50%.

why?
when changing the date period using the date filter, the timestamp generated for the date would have time midday.

fix 🧑‍⚕️🧑‍⚕️
changing the date timestamp to have time of 00:00:00 (which is either midnight at the start of day or midnight at the end, idk)

no more weird offset