simplify the two seperate y-axes into one
using the stack property to stack the bars only not the lines.

this is great because we don't have to manually set the max and min for each axis. and when

setting suggested min and max because if there is pollen values are all at 0, it should be obvious that the chart can go higher than 0 but the value was low. SORRY soooper confusing explanation, definitely easier to show.