bug 🐛🪲: pollen calendar can't handle dates too far apart
when typing in year '2023', the date gets set to '2' at first. the pollen calendar tries to show a date range from 2 to 2024 on the x-axis which causes the component & page to crash 😅😅😁😁

fix ⚒️⚒️: set a min and max that the date filter checks before setting the date.

set min date to be 1/1/2022 because assuming AAPC don't have pollen data before this.

set max date to present because AAPC can't have data for the future 🤔😂