bug fix: when password change fails, backend shouldn't crash 🥲
crashes because of uncaught exception need to use promise resolve & reject in `AWSCognito.auth.service.ts`.

To show a clean error message to user who is trying to change password. Included a try catch in `auth.controller.ts` so that it throws a `BadRequestError` (default error is '500 - Internal Server Error')